refactor(metrics): move record_duplex_coordinate_group + UMI-metrics into shared_metrics
Moves both process_coordinate_group AND update_umi_metrics (the DXM-01/
02/03 guards live in the latter) together, since the former calls the
latter — extracting only one would leave the guards calling back into
code that no longer exists. Drops the metrics: &DuplexMetrics
parameter entirely; the inline path has no such command-struct value.
